requirements.txt
pandas~=2.2.2
PyYAML~=6.0.1
multiprocess~=0.70.16
numpy~=2.0.1
scikit-learn~=1.5.1
deepmerge~=1.1.1
psycopg2~=2.9.9
SQLAlchemy~=2.0.31
networkx~=3.3
tqdm~=4.66.4
pytest~=8.3.1
setuptools~=71.1.0
fastparquet~=2024.5.0
pyarrow~=17.0.0
scipy~=1.13.0
